refactor: move app utils to utils folder
